+++ src/include/sll/memory.h
@@ -21,7 +21,11 @@ __SLL_FUNC __SLL_CHECK_OUTPUT void* sll_reallocate(void* p,sll_size_t sz);
 
 
 
-__SLL_FUNC __SLL_CHECK_OUTPUT void* sll_zero_allocate(sll_size_t sz,sll_size_t c);
+__SLL_FUNC __SLL_CHECK_OUTPUT void* sll_zero_allocate(sll_size_t sz);
+
+
+
+__SLL_FUNC __SLL_CHECK_OUTPUT void* sll_zero_allocate_stack(sll_size_t sz);